feat: migrate storage.spec
